--- app/controllers/editor/edit.js
@@ -1,5 +1,0 @@
-/* eslint-disable ghost/ember/alias-model-in-controller */
-import Controller from '@ember/controller';
-import EditorControllerMixin from 'ghost-admin/mixins/editor-base-controller';
-
-export default Controller.extend(EditorControllerMixin);
