--- app/routes/editor/index.js
@@ -1,8 +1,0 @@
-import Route from '@ember/routing/route';
-
-export default Route.extend({
-    beforeModel() {
-        this._super(...arguments);
-        this.transitionTo('editor.new');
-    }
-});
