+++ app/routes/editor/new.js
@@ -1,31 +1,16 @@
 import AuthenticatedRoute from 'ghost-admin/routes/authenticated';
-import base from 'ghost-admin/mixins/editor-base-route';
 
-export default AuthenticatedRoute.extend(base, {
-    titleToken: 'Editor',
-
+export default AuthenticatedRoute.extend({
     model() {
         return this.get('session.user').then(user => (
-            this.store.createRecord('post', {
-                author: user
-            })
+            this.store.createRecord('post', {author: user})
         ));
     },
 
-    renderTemplate(controller, model) {
-        this.render('editor/edit', {
-            controller,
-            model
-        });
-    },
-
-    actions: {
-        willTransition(transition) {
-            // decorate the transition object so the editor.edit route
-            // knows this was the previous active route
-            transition.data.fromNew = true;
-
-            this._super(...arguments);
-        }
+    // there's no specific controller for this route, instead all editor
+    // handling is done on the editor route/controler
+    setupController(controller, newPost) {
+        let editor = this.controllerFor('editor');
+        editor.setPost(newPost);
     }
 });
