+++ app/templates/components/gh-scheduled-post-countdown.hbs
@@ -1,0 +1,3 @@
+{{#if countdown}}
+    {{yield post countdown}}
+{{/if}}
